+++ src/lib/leaflet-esm.js
@@ -1,15 +1,8 @@
-// Exporting only what we need for our direct usage and for "leaflet.heat"
+// Exporting only what we need for our direct usage
 export {
-  Browser,
-  setOptions,
   divIcon,
-  DomUtil,
   geoJSON,
-  Layer,
   layerGroup,
   map,
   marker,
-  Bounds,
-  point,
-  Tooltip,
 } from 'leaflet/src/Leaflet.js';